+++ h265_converter/__init__.py
@@ -1,4 +1,4 @@
 __author__ = "Draik"
-__date__ = "2024-09-24"
+__date__ = "2024-09-27"
 __status__ = "production"
-__version__ = "0.7.1"
+__version__ = "0.8.0"